requirements.txt
ollama
sentence-transformers
faiss-cpu
langchain
langchain-community
langchain-text-splitters
pypdf
beautifulsoup4
unstructured
numpy
python-dotenv
ollama
sentence-transformers
faiss-cpu
langchain
langchain-community
langchain-text-splitters
pypdf
beautifulsoup4
unstructured
numpy
python-dotenv
datasets
pandas
deepeval